Continue separating pull times logic out of live entry main controller
