+++ app/models/concerns/background_notifiable.rb
@@ -7,15 +7,9 @@ module BackgroundNotifiable
     Turbo::StreamsChannel.broadcast_render_to [event_group, :live_times],
                                               partial: "raw_times/live_times_available",
                                               locals: {
+                                                event_group: event_group,
                                                 unreviewed_count: event_group.raw_times.unreviewed.size,
                                                 unmatched_count: event_group.raw_times.unmatched.size,
                                               }
-    # channel = "event_groups:#{event_group.id}"
-    # message = {
-    #   event: "raw_times_available",
-    #   detail: {unreviewed: event_group.raw_times.unreviewed.size,
-    #            unmatched: event_group.raw_times.unmatched.size}
-    # }
-    # ::ActionCable.server.broadcast(channel, message)
   end
 end
