+++ app/views/live/event_groups/_force_pull_times_button.html.erb
@@ -1,13 +1,15 @@
-<%# locals: (count:) -%>
+<%# locals: (count:, event_group:) -%>
 
 <% count ||= nil %>
 
 <button id="force_pull_times_button"
         class="btn btn-danger d-none"
         data-controller="live-entry--pull-times"
+        data-live-entry--pull-times-event-group-id-value="<%= event_group.id %>"
         data-live-entry--pull-times-force-value="true"
         data-action="keydown@document->live-entry--pull-times#keydown
-                     keyup@document->live-entry--pull-times#keyup"
+                     keyup@document->live-entry--pull-times#keyup
+                     click->live-entry--pull-times#pull"
 >
   Force Pull
   <span class="badge bg-light text-black ms-1"><%= count if count&.positive? %></span>
