+++ app/views/raw_times/_live_times_available.turbo_stream.erb
@@ -1,7 +1,7 @@
-<%# locals: (unreviewed_count:, unmatched_count:) -%>
+<%# locals: (event_group:, unreviewed_count:, unmatched_count:) -%>
 
-<%= turbo_stream.replace "pull_times_button", partial: "live/event_groups/pull_times_button", locals: { count: unreviewed_count } %>
-<%= turbo_stream.replace "force_pull_times_button", partial: "live/event_groups/force_pull_times_button", locals: { count: unmatched_count } %>
+<%= turbo_stream.replace "pull_times_button", partial: "live/event_groups/pull_times_button", locals: { event_group: event_group, count: unreviewed_count } %>
+<%= turbo_stream.replace "force_pull_times_button", partial: "live/event_groups/force_pull_times_button", locals: { event_group: event_group, count: unmatched_count } %>
 <% if unreviewed_count.positive? %>
   <%= turbo_stream.prepend "toasts",
                            partial: "toasts/toast",
